Manually follow EIP-155 in certain circumstances.
Ledger does not seem to handle EIP-155 automatically.  When using a
Ledger, if the block number is at least FORK_BLKNUM = 2675000, one needs
to pre-set the ECDSA signature parameters with r = 0, s = 0, and v =
chainId prior to signing.

The easiest way to handle this on our end is to just branch on the
network type, since mainnet and Ropsten have obviously passed
FORK_BLKNUM.

This is somewhat awkward when dealing with offline transactions, since
we might want to test them on a local network as well.  For now we can
assume that users generating offline transactions wish to later submit
them to mainnet, but the best thing to do is probably to add an
'advanced' tab to offline transaction generation where one can disable
the defaulted-on EIP-155 settings in this case.

See:

See https://github.com/LedgerHQ/ledgerjs/issues/43#issuecomment-366984725

https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md