Fix 'view on Etherscan' links throughout.
At present these always point to mainnet, so are invalid when using
Ropsten, local nodes, etc.

This commit fixes things so that the Etherscan links are disabled, or
are just not visible, when using a network other than Ropsten / mainnet,
and ensures that the links go to the right place when using Ropsten.

Additionally, this fixes a bug where some links were using an old
'disabled' property instead of 'prop-disabled', and so were enabled when
they shouldn't be (including on mainnet).