disable tank for metamask users
Currently, we need a signed transaction to use the gas tank,
because the gas tank demands a signed transaction, both to
determine the sender address, and to send the transaction
directly post-funding, as a theft-deterence mechanism. As
metamask does not support signing a transaction without
broadcasting it, we disable the gas tank for metamask users.