flows: implement sending an invite
Introduces "flows" that can be selected by writing to window.flow.
Supports passing in success and failure callbacks.

Implements such a flow for sending an invite containing a specific
planet to send.

Borrows lots of logic from useInviter. Should be refactored to share it.
Might be worth trying to refactor JustInvite and PassportTransfer to
share their "render progress screen" logic.