release: support approving transfer of lockup
Note that in the "I have both lockups" case this behaves a little funny.
Most of the Release logic tries its best to combine the two lockups,
and so the division there is abstracted away. Approving a transfer
will only do so for _one_ of the lockups, will not do both.

That case should be extremely rare though, so we're probably fine
leaving this as-is for now.