store: improve point sync naming & performance
syncForeignPoint was just an alias for syncDetails. We remove the alias
and use syncDetails directly.

syncOwnedPoints fetches details that are *potentially* relevant if we
own them, but rather unlikely to become useful unless the user
navigates to that point's page. We rename it to syncExtras (though it
still syncs the "base" details as well).

Additionally, and most importantly, we no longer eagerly fetch
the extra details about points when we list them on the Points page.
Instead, we now only get those for the points the user actually
navigates to. This dramatically reduces the number of requests made
when logging into an account with many controlled assets.