store: distinguish SR failure state from loading
As it turns out, trying to fetch star release details for an account may
unexpectedly throw "invalid opcode" errors, if the account in question
does not have any stars in lockup. This in turn would result in the
points list screen staying stuck on "loading...".

We were already handling such exceptions, but were handling them
incorrectly*. We were setting the star release state to `Nothing()`,
which the points list was interpreting as "still loading".

Now, when we encounter any exception, we set the star release state to
`Just(null)`. This means the points list needs to do a little bit more
work to interpret the state, but lets it distinguish between "loading"
and "failed to load".

(* One can argue that "failure to load" does indeed imply we still do not
know what the actual state is, and so we should not set it yet. Even if
that's correct, it's not as practical/convenient as the changes herein.
At least for the time being, we choose "better" over "perfect".)