star release: fix string<->int shenanigans
For linear releases, this was causing us false positives on the "does
this address have a lockup" check. Subsequently calling the withdrawe
limit function for an address without a lockup would result in a
contract-side exception being thrown due to the division by zero. This
exception wasn't getting handled properly, and prevented users with
conditional lockups from actually viewing those in Bridge.

So now we make sure to diligently convert received values into actual
integers, instead of the strings we get them as.

For conditional releases, this wasn't actively causing any problems
(that we were aware of), but we make similar changes there anyway,
for peace of mind.