txn: generalize fake signed transactions
Not all wallets support transaction signing as separate from sending.
Metamask was one such case, we handled it through special logic.
With WalletConnect, we might connect to a wallet that doesn't support
transaction signing either. As such, we generalize Metamask's special
logic.

When the wallet doesn't support transaction signing, we pretend to
succeed, but produce a "fake signed transaction" instead. This has
stubbed out serialization logic, and contains the original transaction
data plus a callback for sending that transaction.

In these cases, the user is informed that their transaction will be
signed by their wallet when sending. The signed transaction fold-out
is not shown.

Worth noting that, for the WalletConnect case, we cannot tell whether
the connected wallet supports transaction signing or not. So we simply
try to do it, and if it fails with METHOD_NOT_SUPPORTED, we know what's
up and proceed with the "fake signed" logic accordingly.