wc: make reticket flow work for walletconnect
The signing and sending callbacks weren't being threaded through.

We get a bit crazy in ResetExecute. Other React state values are
correctly set on-mount, but for some reason the WalletConnect connector
might still be null. We're forced to add it as a dependency to the
useEffect, and write an availability check to match.