metamask: use injected provider for sending txs
Similar to #630, we opt into using the injected provider directly
instead of going through Web3. While reproduction is difficult for this
case, this kind of change worked for the other one, so we go ahead and
apply it here. Certainly can't hurt us more than the bug itself already
does.