keys: during Activation, use new wallet's seed
This fix adds conditional behavior to the MultiKeyFileGenerator.
By default, the keygen uses the logged-in wallet to generate
both halves of the pair.

During the Activation flow, the generator now accepts an optional wallet
arg which is used to pass in the
receiving master ticket wallet. Then,
The MT wallet is used to generate the 2nd
half of the keypair.

This change also renames some variables
for improved readability.

Co-authored-by: willbach <willbach@gmail.com>