proxy: treat L1 points w/ spawn dominion as L2
Before this change, when loading data from the roller API, points that had a dominion of 'spawn' were being treated like legacy L1 points. As such, Azimuth was getting queried, instead of the Roller API. Because of this, users were not seeing updated spawn proxy info in the Bridge UI.

Open question: do we need another Layer indicator? Currently it shows as L1 in the Accounts Dropdown,
but L2 in the ID view

This fixes #874.