keys: address PR feedback, refactor keyfile gen
This commit refactors keyfile generation to use the same base hook `useKeyfileGenerator`  in both the Activation and Home views.

The hook is refactored to accept an array of one or more seeds, and an optional Point instance.

Then, two wrapper hooks are created to reuse the base logic. The difference is in how the seeds are derived for consumption by the base hook.