state: useRoller sets nextBatch in timerStore
On load, Bridge queries the roller for the Config. One of the properties in the response is time to next batch. This value drives the countdown timer in the UI.

Before this change, it was being set in the wrong state context. This change removes the deprecated state context  from rollerStore in favor of timerStore.

Resolves #1004