metamask: ensure expected chain ID
Before this change, it was possible to have Metamask connected to
another EVM chain (e.g. Optimism), or a testnet (e.g. Goerli) while
using Mainnet Bridge. This resolves #1082 by conducting a check for
the expected chain ID in the connected wallet during Login, and
requesting to update it if necessary.